Fix flickering on app load due to theme being loaded later
We read the theme before any of the svelte code is loaded. The benefit
is that it is executed first and there's no flickering if the stored
theme differs from the initial theme that is set in the design system.

The drawback is that this code isn't type-checked and will need to be
updated manually if we change the theme switching code in Svelte.