Fix issue scrolled to the bottom on navigation
If the issue with comments is taller than the viewport then the focus on
the comment reply at the bottom of it would cause the page to be
scrolled to the end when navigating from the issue list. We want to
always open the issue page at the top, this fixes it.