Introduce `update` functions for less error-prone local storage updates
For local storage updates to be persisted it had to be assigned for
example with `storage.value = storage.value`. It was easy to forget this
and it’s not immediately obvious why this is needed. (In fact, eslint
complains.)

We introduce an `update` method for the local storage and an
`updateStoredDraftReview` to encapsulate the necessary logic.